test: move specs into per-package test folders
Colocating specs under src/ mixed test code into what is otherwise
purely publishable source. Move every *.spec.js into a sibling test/
directory that mirrors the src/ layout, matching the convention
already used in flopflip.

Relative specifiers were rewritten by resolving each one against its
old and new directory rather than by string substitution, so plain
imports, dynamic imports and vi.mock() calls all keep pointing at the
same modules.

No configuration changes are needed. Vitest's default include glob is
project-root relative and already covers test/, the package tsconfigs
carry no include list, and tsdown only ever built from src/index.ts.
Every package ships a files allowlist limited to dist/**, so the new
directory stays out of the published tarballs.
