+++ packages/express/test/middleware.spec.js
@@ -1,18 +1,12 @@
-import { createMiddleware as createExpressMetricsMiddleware } from '@promster/express';
 import { createServer as createPrometheusMetricsServer } from '@promster/server';
 import express from 'express';
 import parsePrometheusTextFormat from 'parse-prometheus-text-format';
-import { MockAgent } from 'undici';
-import { afterAll, beforeAll, describe, expect, it } from 'vitest';
+import { afterAll, beforeAll, expect, it } from 'vitest';
 
-import {
-  addObservedAgent,
-  createAgentMetricsExporter,
-  supportedAgentStats,
-} from './agent-metrics';
+import { createMiddleware } from '../src/middleware';
 
-const metricsPort = '1344';
-const appPort = '3023';
+const metricsPort = '1338';
+const appPort = '3001';
 
 const metricsServerUrl = `http://localhost:${metricsPort}`;
 const appServerUrl = `http://localhost:${appPort}`;
@@ -20,7 +14,7 @@ const appServerUrl = `http://localhost:${appPort}`;
 async function startServer() {
   const app = express();
 
-  app.use(createExpressMetricsMiddleware({ app }));
+  app.use(createMiddleware({ app }));
 
   const prometheusMetricsServer = await createPrometheusMetricsServer({
     port: metricsPort,
@@ -63,16 +57,22 @@ async function startServer() {
   };
 }
 
+let app;
 let closeServer;
 
 beforeAll(async () => {
   const startedServer = await startServer();
 
+  app = startedServer.app;
   closeServer = startedServer.close;
 });
 
 afterAll(async () => {
   await closeServer();
+});
+
+it('should expose Prometheus on locals', async () => {
+  expect(app.locals.Prometheus).toBeDefined();
 });
 
 it('should up metric', async () => {
@@ -90,88 +90,66 @@ it('should up metric', async () => {
   );
 });
 
-describe('agent metrics', () => {
-  let mockAgent;
+it('should expose garbage collection metrics', async () => {
+  const response = await fetch(metricsServerUrl);
+  const rawMetrics = await response.text();
 
-  beforeAll(async () => {
-    mockAgent = new MockAgent({ connections: 1 });
-    mockAgent.disableNetConnect();
-
-    const originA = 'http://localhost:9001';
-    const originB = 'http://localhost:9002';
-
-    const agentA = mockAgent.get(originA);
-    const agentB = mockAgent.get(originB);
-
-    agentA
-      .intercept({ path: '/', method: 'GET' })
-      .reply(200, { message: 'ok-a' });
-    agentB
-      .intercept({ path: '/', method: 'POST' })
-      .reply(200, { message: 'ok-b' });
-
-    createAgentMetricsExporter([agentA]);
-    addObservedAgent(agentB);
-
-    const requestToAgentA = await agentA.request({ path: '/', method: 'GET' });
-    await requestToAgentA.body.dump();
-
-    const requestToAgentB = await agentB.request({ path: '/', method: 'POST' });
-    await requestToAgentB.body.dump();
-  });
-
-  afterAll(() => {
-    // Close the mock agent
-    mockAgent.close();
-  });
-
-  it.each(supportedAgentStats)(
-    'for undici agent %s metric',
-    async (statName) => {
-      const response = await fetch(metricsServerUrl);
-      const rawMetrics = await response.text();
-
-      const parsedMetrics = parsePrometheusTextFormat(rawMetrics);
-
-      const nodejsUndiciAgentStats = parsedMetrics.find(
-        (metric) => metric.name === `nodejs_undici_agent_${statName}`,
-      ).metrics;
-
-      const expectedMetrics = {
-        connected: [],
-        free: [],
-        pending: [],
-        queued: [],
-        running: [],
-        size: [],
-      };
-
-      // Use array containment assertion
-      expect(nodejsUndiciAgentStats).toEqual(expectedMetrics[statName]);
-    },
+  const parsedMetrics = parsePrometheusTextFormat(rawMetrics);
+  expect(parsedMetrics).toEqual(
+    expect.arrayContaining([
+      expect.objectContaining({
+        name: 'process_cpu_user_seconds_total',
+      }),
+      expect.objectContaining({
+        name: 'process_cpu_system_seconds_total',
+      }),
+      expect.objectContaining({
+        name: 'process_cpu_seconds_total',
+      }),
+      expect.objectContaining({
+        name: 'process_start_time_seconds',
+      }),
+      expect.objectContaining({
+        name: 'process_resident_memory_bytes',
+      }),
+      expect.objectContaining({
+        name: 'nodejs_eventloop_lag_seconds',
+      }),
+      expect.objectContaining({
+        name: 'nodejs_gc_duration_seconds',
+      }),
+      expect.objectContaining({
+        name: 'nodejs_eventloop_lag_max_seconds',
+      }),
+      expect.objectContaining({
+        name: 'nodejs_eventloop_lag_p50_seconds',
+      }),
+      expect.objectContaining({
+        name: 'nodejs_version_info',
+      }),
+    ]),
   );
-
-  it('should expose the total number of agents', async () => {
-    const response = await fetch(metricsServerUrl);
-    const rawMetrics = await response.text();
-
-    const parsedMetrics = parsePrometheusTextFormat(rawMetrics);
-    const observedAgentsCounter = parsedMetrics.find(
-      (metric) => metric.name === 'nodejs_undici_agents_total',
-    );
-
-    expect(observedAgentsCounter).toBeDefined();
-    expect(observedAgentsCounter.metrics).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          value: '2',
-        }),
-      ]),
-    );
-  });
 });
 
-it('should record the http metrics of requests made', async () => {
+it('should expose http metrics', async () => {
+  const response = await fetch(metricsServerUrl);
+  const rawMetrics = await response.text();
+
+  const parsedMetrics = parsePrometheusTextFormat(rawMetrics);
+
+  expect(parsedMetrics).toEqual(
+    expect.arrayContaining([
+      expect.objectContaining({
+        name: 'http_requests_total',
+      }),
+      expect.objectContaining({
+        name: 'http_request_duration_seconds',
+      }),
+    ]),
+  );
+});
+
+it('should record http metrics', async () => {
   await fetch(appServerUrl);
   const response = await fetch(metricsServerUrl);
   const rawMetrics = await response.text();
