+++ packages/metrics/test/normalizers/status-code.spec.js
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 
-import { normalizeStatusCode } from './status-code';
+import { normalizeStatusCode } from '../../src/normalizers/status-code';
 
 describe('normalizeStatusCode', () => {
   let normalizedStatusCode;
