+++ packages/metrics/test/summary.spec.js
@@ -1,4 +1,4 @@
-vi.mock('./client', () => ({
+vi.mock('../src/client', () => ({
   defaultRegister: {
     metrics: vi.fn(async () => Promise.resolve('metrics')),
     contentType: 'application/test',
@@ -7,8 +7,8 @@ vi.mock('./client', () => ({
 
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { defaultRegister } from './client';
-import { getContentType, getSummary } from './summary';
+import { defaultRegister } from '../src/client';
+import { getContentType, getSummary } from '../src/summary';
 
 describe('getSummary', () => {
   let summary;
