+++ packages/server/test/server.spec.js
@@ -2,7 +2,7 @@ import { createGcMetrics } from '@promster/metrics';
 import parsePrometheusTextFormat from 'parse-prometheus-text-format';
 import { afterAll, beforeAll, expect, it } from 'vitest';
 
-import { createServer } from './server';
+import { createServer } from '../src/server';
 
 const metricsPort = '1342';
 const metricsServerUrl = `http://localhost:${metricsPort}`;
